==> loading/03_QC_datediff.Rmd <==
---
title: "QC datediff features"
author: "Norman Poh"
date: "3 October 2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Init
```{r}
setwd("L:/loading")

library(palab)
library(palabmod)
library(ggplot2)
library(tidyverse)
library(stringr)
library(lubridate)
library(tictoc)
library(hashmap)
library(xgboost)
library(R.utils)
library(ROCR)
library(pROC) # to calculate auc

source("L:/Lib/analyse_res_cv.R")
source("L:/Lib/calculate_eer.R")
source("L:/Lib/config_with_colnames.R")
source("L:/Lib/create_date_diffs.R")
source("L:/Lib/cut_linux.R")
source("L:/Lib/factor2numeric.R")
source("L:/Lib/make_negative.R")
source("L:/Lib/make_positive.R")
source("L:/Lib/make_var_name.R")
source("L:/Lib/mystat.R")
source("L:/Lib/plot_cond_density.R")
source("L:/Lib/remove_inf.R")
source("L:/Lib/remove_na.R")
source("L:/Lib/write_xgb_model.R")

```

## load the data

```{r}
combined = readRDS("D:/Data/Processed/combined_date_complied_rectified_data.rds")
config = read_csv("D:/Data/Processed/combined_date_complied_rectified_config.csv")
results_dir = "D:/Data/Processed/"
mkdirs(results_dir)

```
## Set the post index variables

```{r}
# post index variables
isdiff = str_detect(config$Column, '_diff$')

selected_cols_ = which( config$isdate & config$var_period=="post" )

colnames_ = config$Column[ selected_cols_ ]

colnames_  = c(colnames_, 'disc_dt')
# patients who discontinued
selected_rows_ = which ( !is.na(combined$disc_dt) )

# find out variables
combined_date = combined %>% select(  one_of( colnames_))

date__ <- as.data.frame(  sapply(combined_date, mdy))
date_ = create_date_diffs(date__, index_col = "disc_dt")

stat_ = as.data.frame(t((sapply(date_, mystat))))

todelete = as.data.frame(  sapply(date_, function(x) {x>0}))

# check with a specific example
todelete$post_script_lst[5]
combined$disc_dt[5]
combined$post_script_lst[5]
date_$post_script_lst[5]

# process the todelete var to find non-compliant variables
stat__ =  as.data.frame(sapply(todelete, function(x) {sum(x,  na.rm=TRUE)}) ) 
stat__ = stat__$`sapply(todelete, function(x) {     sum(x, na.rm = TRUE) })`
colnames_ = colnames(date_)
var_non_compliant = colnames_[ stat__!= 0]

var_non_compliant

# ok, delete those
for (i in 1:ncol(date_)) {
  selected_rows_ = which(todelete[,i])
  combined[selected_rows_,colnames_[i]] = NA
  date_[selected_rows_,colnames_[i]] = NA
}
stat_ = as.data.frame(t((sapply(date_, mystat))))
```

## Calculate date diff for the variables that are non-compliant
```{r}

date_ = combined %>% select(  one_of(c(var_non_compliant, 'idx_dt')))

colnames_datediff_var = colnames(date_)

date__ <- as.data.frame(  sapply(date_, mdy))
date_ = create_date_diffs(date__, index_col = "idx_dt")
colnames_original = colnames(date_)
colnames(date_) = paste0(colnames_original, '_diff')

# Let's check them before overwriting -- should be zero or not available
# pick one non-compliant var and check
date_$post_other_outpat_date_lst_diff - combined$post_other_outpat_date_lst_diff

# all good so we write back
colnames_diff = colnames(date_)
for (i in 1:length(colnames_diff)) {
  combined[,colnames_diff[i]] = date_[, colnames_diff[i]] 
}

# saving the data
saveRDS(combined, "D:/Data/Processed/combined_date_complied_data.rds")
saveRDS(config, "D:/Data/Processed/combined_date_complied_config.rds")
write_csv(config, "D:/Data/Processed/combined_date_complied_config.csv")

dim(combined)
dim(config)
```


```{r}
table(combined$post_script_lst_dayssup [ combined$discontinue_flg==1],
  combined$discontinue_flg[ combined$discontinue_flg==1])

table(combined$post_script_lst_dayssup [ combined$discontinue_flg==0],
  combined$discontinue_flg[ combined$discontinue_flg==0])

combined$discontinue_flg = as.factor(combined$discontinue_flg)

ggplot(combined) + aes(post_script_lst_dayssup, group=discontinue_flg,
  fill=discontinue_flg) + geom_histogram(alpha = .2, bins=100) + 
  xlab( config$Description[ config$Column == 'post_script_lst_dayssup'] )

fname = "D:/Data/QC/univar_cond/main_biogen_explore_combined_data_datediff__cond_density__post_script_lst_dayssup_hist.png"

ggsave(fname, plot = last_plot())
```
```{r}
var_non_compliant = c('post_dme_fst','post_script_lst','post_ip_date_lst','post_er_date_lst','post_mgmt_date_lst','post_lab_radio_date_lst','post_other_outpat_date_lst','post_dme_lst','post_ip_date_fst','post_er_date_fst','post_mgmt_date_fst','post_lab_radio_date_fst','post_other_outpat_date_fst')

var_non_compliant = paste0(var_non_compliant,'_diff')

indices = which(config$Column %in% var_non_compliant)
colnames_=colnames(combined)

count_stat = vector('list', length(indices))
for (j in 1:length(indices)) {
  var_index = indices[j]
  print( sprintf("%s", colnames_[var_index]))
  desc_ = config$Description[var_index]
  #print( quantile( as.matrix(combined_data[,indices[j]]), na.rm = TRUE) ) 
  count_stat[[j]] = plot_cond_density(combined, 
    var_index = indices[j], 
    target = combined$discontinue_flg,
    target_label = 'discon',
    description = desc_,
    alpha_cut_off = 0.95, rm.zero = TRUE)
  eer_ = count_stat[[j]]$eer
  fname = sprintf( "D:/Data/QC/univar_cond/main_biogen_explore_combined_data_datediff__cond_density__%s.png",colnames_[indices[j]] )

  ggsave(fname, plot = last_plot())
  
}

eer_ = as.numeric(as.data.frame( lapply(count_stat, function(x) {(x$eer)})) )
pos_ = as.numeric(as.data.frame( lapply(count_stat, function(x) {(x$pos)})) )
neg_ = as.numeric(as.data.frame( lapply(count_stat, function(x) {(x$neg)})) )

eer_table = tibble(Column = colnames_[indices], eer = eer_,`# pos` = pos_, `# neg` = neg_)
write_csv(eer_table, 'D:/Data/QC/num_eer_analysis_complied.csv')


```

## QC the effective_year_diff

```{r}
selected = which(combined$effective_year_diff>0)
combined$

```


## STOP HERE




## check that all pre-index dates are negative
```{r}
if (FALSE) {
  isdate = hashmap(combined_config$Column, combined_config$isdate)
  var_period = hashmap(combined_config$Column, combined_config$var_period)
  
  colnames_ = colnames(date_)
  indices = var_period[[colnames_]]
  
  stat_ = as.data.frame(t((sapply(date_[,indices], mystat))))

  colnames_datediff_var[which(stat_$max>0)]
  View(stat_)
  
  date_[,indices] = make_positive(date_[,indices])
  
  
  # all post-index dates should be negative
  indices = which(period_=="post")
  stat_ = t((sapply(date_[,indices], mystat)))
  View(stat_)
  
  tmp_ = make_negative(date_[,indices])
  stat_ = t(sapply(tmp_, mystat))
  View(stat_)
  
  date_[,indices] = make_negative(date_[,indices])
  
  
  date_$pat_id = combined_data$pat_id
}

saveRDS(date_, "datediff.rds")

# now deal with the config files of the datediff_data
write_csv(date_, 'datediff.csv')

#save the combined_data
dim(combined_data)
dim(date_)

# create the combined file      
combined_data_datediff = cbind(combined_data, date_ %>% select( -one_of('pat_id')) )
saveRDS(combined_data_datediff, 'combined_data_datediff.rds')

combined_config = tibble( Column = colnames(combined_data_datediff ) )
combined_config$Type = var_type[[combined_config$Column]]

selected = str_detect(combined_config$Column, '_diff$')
combined_config$Type[selected] = 'numerical'

#then check the file manually
write_csv(combined_config, 'combined_data_datediff_var_config.csv')

```
