==> loading/05_QC_bivar.Rmd <==
---
title: "Bivariate stat"
author: "Norman Poh"
date: "3 October 2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Init
```{r}
setwd("L:/loading")

library(palab)
library(palabmod)
library(ggplot2)
library(tidyverse)
library(stringr)
library(lubridate)
library(tictoc)
library(hashmap)
library(xgboost)
library(R.utils)
library(ROCR)
library(pROC) # to calculate auc

source("L:/Lib/analyse_res_cv.R")
source("L:/Lib/calculate_eer.R")
source("L:/Lib/config_with_colnames.R")
source("L:/Lib/create_date_diffs.R")
source("L:/Lib/cut_linux.R")
source("L:/Lib/factor2numeric.R")
source("L:/Lib/make_negative.R")
source("L:/Lib/make_positive.R")
source("L:/Lib/make_var_name.R")
source("L:/Lib/mystat.R")
source("L:/Lib/plot_cond_density.R")
source("L:/Lib/remove_inf.R")
source("L:/Lib/remove_na.R")
source("L:/Lib/write_xgb_model.R")

```

```{r}
combined = readRDS("D:/Data/Processed/combined_date_complied_rectified_num_gaba_copay_data.rds")
config = read_csv("D:/Data/Processed/combined_date_complied_rectified_num_gaba_copay_config.csv")
results_dir = "D:/Data/QC_v8/univar"
mkdirs(results_dir)

```

## Univariate stats

```{r}
univar_results <- univariate_stats(input = combined,
  var_config = "D:/Data/Processed/combined_date_complied_rectified_num_gaba_copay_config.csv",
  output_dir = "D:/Data/QC_v8/univar",
  vargt0 = FALSE)

#View(univar_results$cat)

```
## remove variables that are poorly populated
Variable	Problem
pre_concom_gpif_fst8	Variable is 100% missing
pre_concom_gpif_lst8	Variable is 100% missing
pre_medication_fst11	Variable is 100% missing
pre_medication_lst11	Variable is 100% missing
pre_medication_fst13	Variable is 100% missing
pre_medication_lst13	Variable is 100% missing
pre_concom_gpif_fst8_diff	Variable is 100% missing
pre_concom_gpif_lst8_diff	Variable is 100% missing
pre_medication_fst11_diff	Variable is 100% missing
pre_medication_lst11_diff	Variable is 100% missing
pre_medication_fst13_diff	Variable is 100% missing
pre_medication_lst13_diff	Variable is 100% missing
pre_script_fst_diff	Variable is 100% missing
pre_script_lst_dayssup_diff	Variable is 100% missing
post_script_lst_dayssup_diff	Variable is 100% missing
pre_concom_gpif_num8	Variable has only 1 unique value
pre_medication_num11	Variable has only 1 unique value
pre_medication_num13	Variable has only 1 unique value
pre_concom_ndc_fst4_diff	Variable has only 1 unique value
pre_concom_ndc_lst4_diff	Variable has only 1 unique value

pre_concom_ndc_fst4_diff and idx_dt are the same thing

```{r}
#combined = readRDS("D:/Data/Processed/combined_date_complied_rectified_num_data.rds")
#config = read_csv("D:/Data/Processed/combined_date_complied_rectified_num_config.csv")

Description = hashmap(config$Column, config$Description)
config = config_with_colnames(config, colnames(combined))

dim(config)
dim(combined)
sum(config$Column == colnames(combined))

var2remove = c('pre_concom_gpif_fst8','pre_concom_gpif_lst8','pre_medication_fst11','pre_medication_lst11','pre_medication_fst13','pre_medication_lst13','pre_concom_gpif_fst8_diff','pre_concom_gpif_lst8_diff','pre_medication_fst11_diff','pre_medication_lst11_diff','pre_medication_fst13_diff','pre_medication_lst13_diff','pre_script_fst_diff','pre_script_lst_dayssup_diff','post_script_lst_dayssup_diff','pre_concom_gpif_num8','pre_medication_num11','pre_medication_num13','pre_concom_ndc_fst4_diff','pre_concom_ndc_lst4_diff')

length(var2remove)

#remove the variables from data and from the config file for the rest of the QC step
final_config = config %>% filter( ! (config$Column %in% var2remove) )

#make correction
#final_config$Type[ str_detect(final_config$Column,'_diff$') ] = 'numerical'

combined = combined %>% select( -matches(paste(var2remove, collapse="|"))) 

# Check for consistency
nrow(final_config) 
ncol(combined)
sum(colnames(combined) == final_config$Column)

```

## when the outcome is discrete
```{r}
mkdirs("D:/Data/QC_v8/bivar_stats_cat")

# we save the var_config for QC use
write_csv(final_config, "D:/Data/QC_v8/var_config.csv")

bivariate_stats_cat(input = combined, 
                    var_config = "D:/Data/QC_v8/var_config.csv",
                    outcome_var = "discontinue_flg", 
                    output_dir = "D:/Data/QC_v8/bivar_stats_cat",
                    count = 5, 
                    vargt0 = FALSE)
```

## All the post-index variable must have numbers
```{r}
sum(final_config$Column == colnames(combined))

```


## plot all numerical values except those containing flg

```{r}
#debug(plot_cond_density)
#undebug(plot_cond_density)

library(R.utils)

# We would have removed the variables that are 100% missing or with one unique value at this stage
dim(config)
dim(combined)

colnames_ = colnames(combined)

ischar = as.data.frame(sapply(combined, class))
colnames(ischar) = 'class'
ischar$Column = rownames(ischar)
rownames(ischar) = NULL
isnumeric =  ischar$class=='numeric' |ischar$class=='integer'

var_selected = ischar %>% filter(class == 'numeric' |class=='integer') %>% select( one_of('Column'))
 
var2ignore = config$Column[ config$Type=="others" |  
  str_detect(config$Column, "npi|age_group|hce_id|pat_id") ]
    
var_selected = setdiff(var_selected$Column, var2ignore)

#indices  = which(colnames(combined) %in% var_selected)

  
count_stat = vector(mode = "list", length = length(var_selected))

mkdirs("D:/Data/QC_v8/univar_cond")
for (j in 1:length(var_selected)) {
  var_index = var_selected[j]
  print( sprintf("%s", var_index))
  
  #Do not remove zero if the variable is a flag
  if (str_detect(var_index,'flg|der_sex')) {
    rm.zero_=FALSE
  } else {rm.zero_=TRUE}
  
  
  #print( quantile( as.matrix(combined_data[,indices[j]]), na.rm = TRUE) ) 
  count_stat[[j]] = plot_cond_density(combined, 
    var_index = var_index, 
    target = combined$discontinue_flg,
    target_label = 'discon',
    description = Description[[var_index]],
    alpha_cut_off = 0.95, rm.zero = rm.zero_)
  eer_ = count_stat[[j]]$eer
  fname = sprintf( "D:/Data/QC_v8/univar_cond/main_biogen_explore_combined_data_datediff__cond_density__%s.png",var_index )

  ggsave(fname, plot = last_plot())
  
}

eer_ = as.numeric(as.data.frame( lapply(count_stat, function(x) {(x$eer)})) )
pos_ = as.numeric(as.data.frame( lapply(count_stat, function(x) {(x$pos)})) )
neg_ = as.numeric(as.data.frame( lapply(count_stat, function(x) {(x$neg)})) )

eer_table = tibble(Column = var_selected, eer = eer_,`# pos` = pos_, `# neg` = neg_)
write_csv(eer_table, 'D:/Data/QC_v8/num_eer_analysis.csv')

```

## Bivariate stat for numerical variables

```{r}
config_ = final_config %>% filter( Type=="numerical" 
                             | Type =="key" | Column =="discontinue_flg")

write_csv(config_, "D:/Data/QC_v8/var_config_numeric.csv")


data_ = combined %>%   select(matches(paste(config_$Column, collapse="|")))

bivar_stats_y_flag(input = data_, #as.data.frame(numeric(data__)), 
  var_config = "D:/Data/QC_v8/var_config_numeric.csv",
  outcome_var = "discontinue_flg",
  output_dir = "D:/Data/QC_v8/bivar_stats_num")
```
## correlation -- attempt 1 -- not successful
```{r}
config_ = final_config %>% filter( Column %in% var_selected)
write_csv(config_, "D:/Data/QC_v8/correlation/var_config_numeric.csv")

data_ = combined %>%   select( one_of(var_selected ))

univar_results <- univariate_stats(input = data_,
  var_config = "D:/Data/QC_v8/correlation/var_config_numeric.csv",
  output_dir = "D:/Data/QC_v8/correlation",
  vargt0 = FALSE)

#read_csv("D:/Data/QC_v8/correlation/univar_stats_x_num.csv")
NonMissingProp = hashmap(univar_results$numerical$Variable,
  univar_results$numerical$`Non-missing, N`)
varlist = univar_results$numerical$Variable
var2exclude = which(univar_results$numerical$`Non-missing, N` <=50)

varlist = varlist[-var2exclude]

config_ = final_config %>% filter( Column %in% varlist)
write_csv(config_, "D:/Data/QC_v8/correlation/var_config_numeric_corr.csv")

data_ = combined %>%   select( one_of(varlist ))

value_ = correlation(input = data_, 
  var_config = "D:/Data/QC_v8/correlation/var_config_numeric_corr.csv",
  output_dir = "D:/Data/QC_v8/correlation",
  method = "pearson",
  #method = "spearman",
  correlation_warning = TRUE,
  output_csv = TRUE)


out_ = as.data.frame(sort(unlist(lapply(data_, function(x) sum(!is.na(x))))))
colnames(out_)="value"
countMap = hashmap(rownames(out_), out_$value)

```
## correlation -- attempt 2
```{r}
Description = hashmap(config$Column, config$Description)

# detect the variable class by type
ischar = as.data.frame(sapply(combined, class))
colnames(ischar) = 'class'
ischar$Column = rownames(ischar)
rownames(ischar) = NULL

var_selected = ischar %>% filter(class == 'numeric' |class=='integer') %>% select( one_of('Column'))

# get the selected variable ==> var_selected
var_selected = ischar %>% filter(class == 'numeric' |class=='integer') %>% select( one_of('Column'))
 
var2ignore = config$Column[ config$Type=="others" |  
  str_detect(config$Column, "npi|age_group|hce_id|pat_id") ]
    
var_selected = setdiff(var_selected$Column, var2ignore)

data_ = combined %>%   select( one_of(var_selected ))

stat_ = as.data.frame(sort(unlist(lapply(data_, function(x) sum(!is.na(x))))))
colnames(stat_)="value"; stat_$Column = rownames(stat_); rownames(stat_)=NULL

#combi_ = pracma::nchoosek(length(stat_$Column), 2)
colnames_ = colnames(data_)
combi_ = combn(length(colnames_), 2)
combi_ = t(combi_) 

# define a helper function to count concurrent occurence of a pair of variables given their indices
is.na_pair <- function(pair) {
  selected = !is.na( data_[,pair[1]]) & !is.na( data_[,pair[2]])
  return(sum(selected))
}

#apply row-by-row
out_ = as.data.frame(unlist(apply(combi_, 1,is.na_pair)))
colnames(out_) = 'Count'

#create the table
table_ = tibble(V1 = combi_[,1], V2 = combi_[,2] )
table_$Count = out_$Count
table_$`Completeness` = table_$Count/nrow(data_)
table_$Description_V1 = Description[[colnames_[combi_[,1]]]]
table_$Description_V2 = Description[[colnames_[combi_[,2]]]]

table__ = table_ %>% arrange(desc(Count)) %>% filter( Count >=5)

# Compute correlation

correlation <- function(pair) {
  selected = !is.na( data_[,pair[1]]) & !is.na( data_[,pair[2]])
  value_pearson = cor.test(data_[selected,pair[1]], data_[selected,pair[2]],
           method = 'pearson')
  value_spearman = cor.test(data_[selected,pair[1]], data_[selected,pair[2]],
           method = 'spearman')
  return( list(pearson_corr = value_pearson$estimate,
     pearson_p.value = value_pearson$p.value,
     spearman_rho = value_spearman$estimate,
     spearman_p.value = value_pearson$p.value
  ))
}

combi__ = cbind(table__$V1,table__$V2)
dim(combi__)
tictoc::tic()
#out_ = data.table::rbindlist(as.matrix(apply(combi__[1:100,], 1,correlation)))

out_ = data.table::rbindlist(as.matrix(apply(combi__, 1,correlation)))
tictoc::toc()
```


## Further filter out variables
```{r}
eer_table = read_csv('D:/Data/QC_v8/num_eer_analysis.csv')

# windows2unix_path <- function (windows_path) {
#   return(str_replace(windows_path,'\\','/'))
# }
  
table_ = eer_table %>% filter( eer <=0.04 )
write_csv(table_,'D:/Data/QC_v8/var_removed_due_to_low_eer.csv')

View(eer_table)
table_ = eer_table %>% filter( eer >0.04 )

table__ = table_ %>% filter( `# pos`<=5 )
write_csv(table__,'D:/Data/QC_v8/var_removed_due_to_npos_ls_05.csv')

table__ = eer_table %>% filter( eer >0.04 ) %>% filter( `# pos` > 5 & `# pos`<=10 )
write_csv(table__,'D:/Data/QC_v8/var_removed_due_to_npos_ls_10.csv')

```

## remaining variables
```{r}

table.remaining = eer_table %>% filter( eer >0.04 ) %>% filter( `# pos`>10 ) 
write_csv(table.remaining,'D:/Data/QC_v8/var_remained.csv')

```
